requirements.txt
asgiref==3.5.2
Django==4.1
django-ninja==0.19.1
gunicorn==20.1.0
orjson==3.7.12
pydantic==1.9.2
sqlparse==0.4.2
typing_extensions==4.3.0
whitenoise==6.2.0
